Move listener to presentation layer.
