+++ src/main/java/oripa/gui/view/main/MainFrame.java
@@ -45,7 +45,6 @@ import org.slf4j.LoggerFactory;
 import oripa.application.main.DataFileAccess;
 import oripa.application.main.IniFileAccess;
 import oripa.application.main.PaintContextModification;
-import oripa.application.main.UnselectAllItemsActionListener;
 import oripa.appstate.StatePopper;
 import oripa.doc.Doc;
 import oripa.domain.cptool.Painter;
@@ -65,6 +64,7 @@ import oripa.gui.presenter.creasepattern.DeleteSelectedLinesActionListener;
 import oripa.gui.presenter.creasepattern.EditMode;
 import oripa.gui.presenter.creasepattern.MouseActionHolder;
 import oripa.gui.presenter.creasepattern.SelectAllLineActionListener;
+import oripa.gui.presenter.creasepattern.UnselectAllItemsActionListener;
 import oripa.gui.viewsetting.ViewScreenUpdater;
 import oripa.gui.viewsetting.main.MainFrameSetting;
 import oripa.gui.viewsetting.main.MainScreenSetting;
